[FEAT] configure Spring Security with JWT authentication and admin protection
