+++ src/main/kotlin/com/hbd/book_be/security/JwtAuthenticationFilter.kt
@@ -1,0 +1,36 @@
+package com.hbd.book_be.security
+
+import jakarta.servlet.FilterChain
+import jakarta.servlet.http.HttpServletRequest
+import jakarta.servlet.http.HttpServletResponse
+import org.springframework.security.core.context.SecurityContextHolder
+import org.springframework.stereotype.Component
+import org.springframework.web.filter.OncePerRequestFilter
+
+@Component
+class JwtAuthenticationFilter(
+    private val jwtTokenProvider: JwtTokenProvider
+) : OncePerRequestFilter() {
+    
+    override fun doFilterInternal(
+        request: HttpServletRequest, 
+        response: HttpServletResponse, 
+        filterChain: FilterChain
+    ) {
+        val token = extractTokenFromRequest(request)
+        
+        if (token != null && jwtTokenProvider.validateToken(token)) {
+            val authentication = jwtTokenProvider.getAuthentication(token)
+            SecurityContextHolder.getContext().authentication = authentication
+        }
+        
+        filterChain.doFilter(request, response)
+    }
+    
+    private fun extractTokenFromRequest(request: HttpServletRequest): String? {
+        val bearerToken = request.getHeader("Authorization")
+        return if (bearerToken != null && bearerToken.startsWith("Bearer ")) {
+            bearerToken.substring(7)
+        } else null
+    }
+}
